removed all types directories
