+++ src/structures/CommandManager.ts
@@ -3,14 +3,13 @@ import { Guild, Interaction, Message } from "discord.js";
 import { InputParameter, ObjectID, TargetID } from "./Parameter";
 import { CommandNotFound } from "../errors";
 import { applicationState } from "../state";
-import { ChatCommand } from "../commands/ChatCommand";
-import { ContextMenuCommand } from "../commands/ContextMenuCommand";
-import { Commands, CommandInit, CommandRegExps, CommandType } from "../commands/types/commands";
+import { ChatCommand, ChatCommandInit } from "../commands/ChatCommand";
+import { ContextMenuCommand, ContextMenuCommandInit } from "../commands/ContextMenuCommand";
+import { Commands, CommandInit, CommandRegExps, CommandType } from "../commands/commandsTypes";
 import { APICommandObject, CommandPermission, RegisteredCommandObject, APICommandType } from "./apiTypes";
 import { Bot } from "./Bot";
 import { SubCommand } from "../commands/SubCommand";
 import { SubCommandGroup } from "../commands/SubCommandGroup";
-import { ChatCommandInit, ContextMenuCommandInit } from "../commands/types/InitOptions";
 import { HelpMessageParams } from "../commands/Help";
 import { HelpMessage } from "../commands/Help";
 import { PrefixManager } from "./PrefixManager";
